Exceptions + tests for CarService
Done all exceptions and tests for CarService, still no for StringUtils.

+++ src/main/java/pl/edu/pjatk/LAB_2/service/CarService.java
@@ -1,66 +1,98 @@
 package pl.edu.pjatk.LAB_2.service;
 
 import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.RequestBody;
+import pl.edu.pjatk.LAB_2.exceptions.CarAlreadyExistsException;
+import pl.edu.pjatk.LAB_2.exceptions.CarNotFoundExceptions;
+import pl.edu.pjatk.LAB_2.exceptions.CarWrongDataInputException;
 import pl.edu.pjatk.LAB_2.model.Car;
 import pl.edu.pjatk.LAB_2.repository.CarRepository;
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Component
 public class CarService {
-    List<Car> carList = new ArrayList<>();
 
-    private StringUtilsService stringUtilsService;
-    private CarRepository repository;
+    private final StringUtilsService stringUtilsService;
+    private final CarRepository repository;
 
     public CarService(CarRepository repository, StringUtilsService stringUtilsService) {
         this.repository = repository;
         this.stringUtilsService = stringUtilsService;
-        this.repository.save(new Car("RS6", "black"));
-        this.repository.save(new Car("RS7", "red"));
-        this.repository.save(new Car("R8", "yellow"));
+
+        add(new Car("RS6", "black"));
+        add(new Car("RS7", "red"));
+        add(new Car("R8", "yellow"));
     }
 
     public List<Car> getCarByModel(String model) {
-     return this.repository.findByModel(model);
+        if(model == null || model.isEmpty()) throw new CarWrongDataInputException();
+        List<Car> cars = repository.findByModel(model);
+        if(cars.isEmpty()) throw new CarNotFoundExceptions();
+        return cars;
     }
+
     public List<Car> getCarList() {
-        List<Car> cars = (List<Car>) this.repository.findAll(); ;
+        List<Car> cars = (List<Car>) this.repository.findAll();
+        if(cars.isEmpty()) throw new CarNotFoundExceptions();
         for (Car car : cars) {
             car.setModel(stringUtilsService.toProperCase(car.getModel()));
             car.setColor(stringUtilsService.toProperCase(car.getColor()));
         }
         return cars;
     }
-    public void add(@RequestBody Car car){
+
+    public void checkIfCarExists(Car car) {
+        List<Car> carsBySum = this.repository.findByCharToIntSum(car.getCharToIntSum());
+        if (!carsBySum.isEmpty()) {
+            throw new CarAlreadyExistsException();
+        }
+    }
+
+    public void add(Car car) {
+        carWrongDataInputException(car);
+        checkIfCarExists(car);
         car.setModel(stringUtilsService.toProperCase(car.getModel()));
         car.setColor(stringUtilsService.toProperCase(car.getColor()));
         this.repository.save(car);
     }
-    public Optional<Car> getCar(Long id) {
+
+    public Car getCar(Long id) {
         Optional<Car> carOptional = this.repository.findById(id);
-        carOptional.ifPresent(car -> {
-            car.setModel(stringUtilsService.toProperCase(car.getModel()));
-            car.setColor(stringUtilsService.toProperCase(car.getColor()));
-        });
-        return carOptional;
+        if(carOptional.isEmpty()) throw new CarNotFoundExceptions();
+
+        Car car = carOptional.get();
+        car.setModel(stringUtilsService.toProperCase(car.getModel()));
+        car.setColor(stringUtilsService.toProperCase(car.getColor()));
+        return car;
     }
+
     public void delete(Long id){
+        Optional<Car> carOptional = this.repository.findById(id);
+        if(carOptional.isEmpty()) throw new CarNotFoundExceptions();
+
         this.repository.deleteById(id);
     }
-    public void update(Long id, Car car){
-        Optional<Car> existingCarOptional = repository.findById(car.getId());
 
-        if(existingCarOptional.isPresent()){
-            Car existingCar = existingCarOptional.get();
-            existingCar.setModel(stringUtilsService.toProperCase(car.getModel()));
-            existingCar.setColor(stringUtilsService.toProperCase(car.getColor()));
-            this.repository.save(existingCar);
-        }else {
-            throw new RuntimeException("Samochód o ID " + car.getId() + " nie istnieje.");
-        }
+    public void update(Long id, Car car){
+        carWrongDataInputException(car);
+
+        Optional<Car> carOptional = this.repository.findById(id);
+
+        if(carOptional.isEmpty()) throw new CarNotFoundExceptions();
+
+        Car carToUpdate = carOptional.get();
+        checkIfCarExists(car);
+
+        carToUpdate.setModel(stringUtilsService.toProperCase(car.getModel()));
+        carToUpdate.setColor(stringUtilsService.toProperCase(car.getColor()));
+        carToUpdate.setCharToIntSum(car.getCharToIntSum());
+
+        this.repository.save(carToUpdate);
+    }
+
+    public void carWrongDataInputException(Car car){
+        if(car.getModel() == null || car.getModel().isEmpty()) throw new CarWrongDataInputException();
+        if(car.getColor() == null || car.getColor().isEmpty()) throw new CarWrongDataInputException();
     }
 }
+
